Update Passwords Module and other minor tweaks.
Update Passwords Module to be quicker for the user.
Changed Tab keybind to F1 to allow Tab to be used to navigate like normal.
Added Timer Group back to the main screen. This still needs functionality to make it actually work.
More work on debug output. It's still too slow so it is not enabled yet.